requirements.txt
# --------- pytorch --------- #
torch>=1.8.0
torchvision>=0.9.1
pytorch-lightning==1.6.5
torchmetrics==0.9.2

# --------- hydra --------- #
hydra-core==1.2.0
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
# wandb
# neptune-client
# mlflow
# comet-ml

# --------- others --------- #
pyrootutils     # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
sh              # for running bash commands in some tests
